Separate header and footer from VueScreener, load these separately
