+++ src/stories/1-basic-usage/3-array-of-mixed-objects.story.vue
@@ -1,0 +1,14 @@
+<template>
+  <Story title="3. Array of mixed objects" group="basic-usage" source="-">
+    <ScreenerHeader :screener="screener" />
+    <VueScreener :screener="screener" />
+    <ScreenerFooter :screener="screener" />
+  </Story>
+</template>
+
+<script lang="ts" setup>
+import { VueScreener, Header as ScreenerHeader, Footer as ScreenerFooter } from '../../index'
+import mixedObjectsData from '../../fixtures/mix-objects-data.json'
+import { useScreener } from '../../hooks/use-screener'
+const screener = useScreener(mixedObjectsData, { title: 'Array of mixed objects' })
+</script>
