+++ src/stories/1-basic-usage/6-custom-title.story.vue
@@ -1,0 +1,14 @@
+<template>
+  <Story title="6. Custom title" group="basic-usage" source="-">
+    <ScreenerHeader :screener="screener" />
+    <VueScreener :screener="screener" />
+    <ScreenerFooter :screener="screener" />
+  </Story>
+</template>
+
+<script lang="ts" setup>
+import { VueScreener, Header as ScreenerHeader, Footer as ScreenerFooter } from '../../index'
+import baseData from '../../fixtures/data.json'
+import { useScreener } from '../../hooks/use-screener'
+const screener = useScreener(baseData, { title: 'Custom title' })
+</script>
