+++ src/stories/2-configure-columns/1-control-widths.story.vue
@@ -1,0 +1,61 @@
+<template>
+  <Story title="1. Control widths" group="configure-columns" source="-">
+    <ScreenerHeader :screener="screener1" />
+    <VueScreener :screener="screener1" />
+    <ScreenerFooter :screener="screener1" />
+    <br />
+    <ScreenerHeader :screener="screener2" />
+    <VueScreener :screener="screener2" />
+    <ScreenerFooter :screener="screener2" />
+    <br />
+    <ScreenerHeader :screener="screener3" />
+    <VueScreener :screener="screener3" />
+    <ScreenerFooter :screener="screener3" />
+  </Story>
+</template>
+
+<script lang="ts" setup>
+import { VueScreener, Header as ScreenerHeader, Footer as ScreenerFooter } from '../../index'
+import baseData from '../../fixtures/data.json'
+import primitivesData from '../../fixtures/primitives-data.json'
+import mixedObjectsData from '../../fixtures/mix-objects-data.json'
+import { useScreener } from '../../hooks/use-screener'
+
+const screener1 = useScreener(baseData, {
+  title: 'Results',
+  config: {
+    id: { field: 'id', width: '50px' },
+    first_name: { field: 'first_name', width: '150px' },
+    last_name: { field: 'last_name', width: '150px' },
+    full_name: { field: 'full_name', width: '150px' },
+    email: { field: 'email', width: '200px' },
+    gender: { field: 'gender', width: '100px' },
+    ip_address: { field: 'ip_address', width: 'minmax(150px, 1fr)' },
+  },
+})
+
+const screener2 = useScreener(primitivesData, {
+  title: 'Results',
+  config: {
+    0: { field: 0, width: '50px' },
+    1: { field: 1, width: '100px' },
+    2: { field: 2, width: '100px' },
+    3: { field: 3, width: '1fr' },
+    4: { field: 4, width: '100px' },
+    5: { field: 5, width: '150px' },
+  },
+})
+
+const screener3 = useScreener(mixedObjectsData, {
+  title: 'Results',
+  pick: ['id', 'type', 'name', 'address', 'country', 'flag_colours'],
+  config: {
+    type: { field: 'type', width: '75px' },
+    name: { field: 'name', width: '100px' },
+    id: { field: 'id', width: '50px' },
+    address: { field: 'address', width: 'minmax(200px, 1fr)' },
+    country: { field: 'country', width: '100px' },
+    flag_colours: { field: 'flag_colours', width: '200px' },
+  },
+})
+</script>
