+++ src/stories/2-configure-columns/7-pick-omit-fields.story.vue
@@ -1,0 +1,19 @@
+<template>
+  <Story title="7. Pick + omit fields" group="configure-columns" source="-">
+    <ScreenerHeader :screener="screener" />
+    <VueScreener :screener="screener" />
+    <ScreenerFooter :screener="screener" />
+  </Story>
+</template>
+
+<script lang="ts" setup>
+import { VueScreener, Header as ScreenerHeader, Footer as ScreenerFooter } from '../../index'
+import baseData from '../../fixtures/data.json'
+import { useScreener } from '../../hooks/use-screener'
+
+const screener = useScreener(baseData, {
+  title: 'Results',
+  pick: ['id', 'first_name', 'last_name'],
+  omit: ['id'],
+})
+</script>
